requirements.txt
gspread~=5.7.2
pandas~=1.5.3
APScheduler~=3.6.3
tradingview_ta~=3.3.0
python_telegram_bot~=13.15